Get rid of RuntimeWarning: 'beniget.beniget' found in sys.modules after import of package 'beniget'...
Use the adequate __main__.py file to store beniget command line
interface.

As a result one should now invoke beniget through `python -m beniget file.py`
